requirements.txt
adlinktech-datariver>=1.4.1